% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nombres}
\alias{nombres}
\title{Nombre de Personas.}
\format{Un data frame con 3 variables: \code{anio}, \code{nombre},
  \code{cantidad}}
\usage{
nombres
}
\description{
Nombres registrados en el Registro Nacional de las Personas agrupados por año.
La información surge de los puestos de toma de trámites del DNI en todo el país,
incluyendo argentinos y extranjeros nacionalizados. (1922-2015)
}
\keyword{datasets}
